+++ Medical Inventory/Data/IRepository/ICompanyRepository.cs
@@ -4,7 +4,7 @@ namespace Medical_Inventory.Data.IRepository;
 
 public interface ICompanyRepository
 {
-    Task<Company?>? GetFirstOrDefault(int? id);
+    Task<Company?>? GetFirstOrDefault(long? id);
     Task<IEnumerable<Company>?>? GetAll();
     Task Add(Company entity);
     Task Update(Company entity);
