+++ Medical Inventory/Data/IRepository/IGenericRepository.cs
@@ -4,7 +4,7 @@ namespace Medical_Inventory.Data.IRepository;
 
 public interface IGenericRepository
 {
-    Task<Generic?>? GetFirstOrDefault(int? id);
+    Task<Generic?>? GetFirstOrDefault(long? id);
     Task<IEnumerable<Generic>?>? GetAll();
     Task Add(Generic entity);
     Task Update(Generic entity);
